requirements.txt
alembic==1.8.0
asn1crypto==1.5.1
attrs==21.4.0
certifi==2022.5.18.1
click==8.1.3
commonmark==0.9.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pg8000==1.26.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.1
python-dotenv==0.20.0
scramp==1.4.1
SQLAlchemy==1.4.35
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
